Clean things up and add build checker. (#22)
* Clean things up and add build checker.

* Update R version.

Match the full build in case that's lower for a reason.

* Maybe versions will help?

* Install system dependencies.

* Can this work on Mac?

* Maybe build on Ubuntu?

* Wait 16.04 isn't available.

* We'll try with 18.04 before giving up.

* Revert to macOS

This is what everyone uses so I'm leaving it like this. We'll need to get rid of crrry to get this to work. Maybe I'll deal with that later.

* revert r-blocks to formatted text

* remove {crrry} dependency

* Revert to known working versions of r-lib actions.

* Explicitly require xfun.

I think our cache might be broken, so I'm not sure this will help, but it's what specifically broke.

* Remove R 4.0.5 req.

Maybe that'll bust the cache or whatever is failing now.

Co-authored-by: Russ Hyde <russ.hyde.data@gmail.com>